Run clippy

+++ src/monster_ai_system.rs
@@ -56,8 +56,8 @@ impl<'a> System<'a> for MonsterAI {
                     map.blocked[idx] = false;
 
                     // New position which is calculated by A*
-                    pos.x = path.steps[1] as i32 % map.width as i32;
-                    pos.y = path.steps[1] as i32 / map.width as i32;
+                    pos.x = path.steps[1] as i32 % map.width;
+                    pos.y = path.steps[1] as i32 / map.width;
 
                     idx = map.map_index(pos.x, pos.y);
